requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
sse-starlette==1.6.5
starlette==0.27.0
httpx==0.25.2
aiofiles==23.2.1